requirements.txt
pytest-playwright==0.4.2
PyYAML==6.0.1
pytest==7.4.0
pytest-bdd==6.1.1